+++ types/meson.build
@@ -27,7 +27,12 @@ wlr_files += files(
 	'xdg_shell/wlr_xdg_shell.c',
 	'xdg_shell/wlr_xdg_surface.c',
 	'xdg_shell/wlr_xdg_toplevel.c',
-	'wlr_buffer.c',
+	'buffer/buffer.c',
+	'buffer/client.c',
+	'buffer/dmabuf.c',
+	'buffer/readonly_data.c',
+	'buffer/resource.c',
+	'buffer/shm_client.c',
 	'wlr_compositor.c',
 	'wlr_cursor.c',
 	'wlr_damage_ring.c',
